+++ AutoGestion/Funciones.h
@@ -9,5 +9,5 @@ public:
 	bool esDecimal(std::string s);
 	bool esNumero(std::string s);
     int inputNumero(int x, int y);
-    int inputDecimal(int x, int y);
+    float inputDecimal(int x, int y);
 };
